SB-715: some `LayoutProvider` methods has been extracted to `AbstractLayoutProvider` with common implementation
