+++ strongbox-storage/strongbox-storage-api/src/main/java/org/carlspring/strongbox/providers/layout/Maven2LayoutProvider.java
@@ -1,22 +1,15 @@
 package org.carlspring.strongbox.providers.layout;
 
 import java.io.IOException;
-import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.security.NoSuchAlgorithmException;
-import java.util.Map;
 
 import javax.annotation.PostConstruct;
 
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.repository.metadata.Metadata;
 import org.carlspring.maven.commons.util.ArtifactUtils;
-import org.carlspring.strongbox.artifact.coordinates.ArtifactCoordinates;
 import org.carlspring.strongbox.artifact.coordinates.MavenArtifactCoordinates;
-import org.carlspring.strongbox.io.ArtifactPath;
 import org.carlspring.strongbox.io.RepositoryFileSystemProvider;
 import org.carlspring.strongbox.io.RepositoryPath;
 import org.carlspring.strongbox.storage.Storage;
@@ -85,174 +78,31 @@ public class Maven2LayoutProvider extends AbstractLayoutProvider<MavenArtifactCo
         return ArtifactUtils.isChecksum(path);
     }
 
-    @Override
-    public void copy(String srcStorageId,
-                     String srcRepositoryId,
-                     String destStorageId,
-                     String destRepositoryId,
-                     String path)
-            throws IOException
+    protected void doDeletePath(RepositoryPath repositoryPath, boolean force) throws IOException
     {
-        // TODO: Implement
-    }
+        RepositoryPath md5Path = repositoryPath.resolveSibling(repositoryPath.getFileName() + ".md5");
+        RepositoryPath sha1Path = repositoryPath.resolveSibling(repositoryPath.getFileName() + ".sha1");
 
-    @Override
-    public void move(String srcStorageId,
-                     String srcRepositoryId,
-                     String destStorageId,
-                     String destRepositoryId,
-                     String path)
-            throws IOException
-    {
-        // TODO: Implement
-    }
-
-    @Override
-    public void delete(String storageId,
-                       String repositoryId,
-                       String path,
-                       boolean force)
-            throws IOException
-    {
-        Storage storage = getConfiguration().getStorage(storageId);
-        Repository repository = storage.getRepository(repositoryId);
+        Files.delete(repositoryPath);
+        Files.deleteIfExists(md5Path);
+        Files.deleteIfExists(sha1Path);
         
-        RepositoryPath repositoryBasePath = resolve(repository);
-        RepositoryPath repositoryPath = repositoryBasePath.resolve(path);
-
-        logger.debug("Checking in " + storageId + ":" + repositoryId + "(" + path + ")...");
-        if (!Files.exists(repositoryPath))
-        {
-            logger.warn(String.format("Path not found: path-[%s]", repositoryPath));
-            return;
-        }
-
+        Repository repository = repositoryPath.getFileSystem().getRepository();
         RepositoryFileSystemProvider provider = getProvider(repositoryPath);
-        if (!Files.isDirectory(repositoryPath)){
-            RepositoryPath md5Path = repositoryPath.resolveSibling(repositoryPath.getFileName() + ".md5");
-            RepositoryPath sha1Path = repositoryPath.resolveSibling(repositoryPath.getFileName() + ".sha1");
-            
-            Files.delete(repositoryPath);
-            Files.deleteIfExists(md5Path);
-            Files.deleteIfExists(sha1Path);
-        
-            if (force && repository.allowsForceDeletion()){
-                provider.deleteTrash(repositoryPath);
-                provider.deleteTrash(md5Path);
-                provider.deleteTrash(sha1Path);
-            }
-        } else {
-            Files.walkFileTree(repositoryPath, new SimpleFileVisitor<Path>()
-            {
-                @Override
-                public FileVisitResult visitFile(Path file,
-                                                 BasicFileAttributes attrs)
-                    throws IOException
-                {
-                    Files.delete(file);
-                    return FileVisitResult.CONTINUE;
-                }
-
-                @Override
-                public FileVisitResult postVisitDirectory(Path dir,
-                                                          IOException exc)
-                    throws IOException
-                {
-                    Files.delete(dir);
-                    return FileVisitResult.CONTINUE;
-                }
-            });
-        }
-        
-        logger.debug("Removed /" + repositoryId + "/" + path);
-    }
-
-    @Override
-    public void deleteTrash(String storageId, String repositoryId)
-            throws IOException
-    {
-        logger.debug("Emptying trash for repositoryId " + repositoryId + "...");
-        Storage storage = getConfiguration().getStorage(storageId);
-        Repository repository = storage.getRepository(repositoryId);
-        RepositoryPath path = resolve(repository);
-        RepositoryFileSystemProvider provider = (RepositoryFileSystemProvider)path.getFileSystem().provider();
-        
-        provider.deleteTrash(path);
-    }
-
-    //TODO: reimplement with Path
-    @Override
-    public void deleteTrash()
-            throws IOException
-    {
-        for (Map.Entry entry : getConfiguration().getStorages().entrySet())
+        if (force && repository.allowsForceDeletion())
         {
-            Storage storage = (Storage) entry.getValue();
-
-            final Map<String, Repository> repositories = storage.getRepositories();
-            for (Repository repository : repositories.values())
-            {
-                if (!repository.allowsDeletion())
-                {
-                    logger.warn("Repository " + repository.getId() + " does not support removal of trash.");
-                }
-                deleteTrash(storage.getId(), repository.getId());
-            }
+            provider.deleteTrash(repositoryPath);
+            provider.deleteTrash(md5Path);
+            provider.deleteTrash(sha1Path);
         }
     }
-
-    @Override
-    public void undelete(String storageId,
-                         String repositoryId,
-                         String path)
-        throws IOException
-    {
-        logger.debug(String.format("Attempting to restore: storageId-[%s]; repoId-[%s]; path-[%s]; ", storageId,
-                                   repositoryId, path));
-        ArtifactPath artifactPath = resolve(storageId, repositoryId, path);
-
-        RepositoryFileSystemProvider provider = getProvider(artifactPath);
-        provider.restoreTrash(artifactPath);
-    }
-
-    @Override
-    public void undeleteTrash(String storageId, String repositoryId)
-            throws IOException
-    {
-        Storage storage = getConfiguration().getStorage(storageId);
-        Repository repository = storage.getRepository(repositoryId);
-
-        logger.debug("Restoring all artifacts from the trash of " + storageId + ":" + repository.getId() + "...");
-        if (!repository.isTrashEnabled())
-        {
-            logger.warn("Repository " + repository.getId() + " does not support removal of trash.");
-        }
-        
-        RepositoryPath path = resolve(repository);
-        getProvider(path).restoreTrash(path);
-    }
-
-    @Override
-    public void undeleteTrash()
-            throws IOException
-    {
-        for (Map.Entry entry : getConfiguration().getStorages().entrySet())
-        {
-            Storage storage = (Storage) entry.getValue();
-
-            final Map<String, Repository> repositories = storage.getRepositories();
-            for (Repository repository : repositories.values())
-            {
-                undeleteTrash(storage.getId(), repository.getId());
-            }
-        }
-    }
+    
     
     @Override
     public void deleteMetadata(String storageId,
                                String repositoryId,
                                String metadataPath)
-            throws IOException
+        throws IOException
     {
         // TODO: Further untangle the relationships of this so that the code below can be uncommented:
 
@@ -263,14 +113,14 @@ public class Maven2LayoutProvider extends AbstractLayoutProvider<MavenArtifactCo
         {
             return;
         }
-        
+
         try
         {
             String version = repositoryPath.getFileName().toString();
             java.nio.file.Path path = repositoryPath.getParent();
             Metadata metadata = mavenMetadataManager.readMetadata(path);
             if (metadata != null && metadata.getVersioning() != null
-                && metadata.getVersioning().getVersions().contains(version))
+                    && metadata.getVersioning().getVersions().contains(version))
             {
                 metadata.getVersioning().getVersions().remove(version);
                 mavenMetadataManager.storeMetadata(path, null, metadata, MetadataType.ARTIFACT_ROOT_LEVEL);
@@ -281,31 +131,4 @@ public class Maven2LayoutProvider extends AbstractLayoutProvider<MavenArtifactCo
             // We won't do anything in this case because it doesn't have an impact to the deletion
         }
     }
-
-    @Override
-    public boolean contains(String storageId, String repositoryId, String path)
-            throws IOException
-    {
-        ArtifactPath artifactPath = resolve(storageId, repositoryId, path);
-        return Files.exists(artifactPath);
-    }
-
-    @Override
-    public boolean containsArtifact(Repository repository,
-                                    ArtifactCoordinates coordinates)
-        throws IOException
-    {
-        ArtifactPath artifactPath = resolve(repository, coordinates);
-        return Files.exists(artifactPath);
-    }
-
-    @Override
-    public boolean containsPath(Repository repository, String path)
-            throws IOException
-    {
-        RepositoryPath repositoryPath = resolve(repository);
-
-        return Files.exists(repositoryPath.resolve(path));
-    }
-
 }
