+++ strongbox-storage/strongbox-storage-api/src/main/java/org/carlspring/strongbox/providers/layout/P2LayoutProvider.java
@@ -1,9 +1,12 @@
 package org.carlspring.strongbox.providers.layout;
 
 import java.io.IOException;
+import java.nio.file.Files;
 
 import org.carlspring.strongbox.artifact.coordinates.ArtifactCoordinates;
 import org.carlspring.strongbox.artifact.coordinates.P2ArtifactCoordinates;
+import org.carlspring.strongbox.io.RepositoryFileSystemProvider;
+import org.carlspring.strongbox.io.RepositoryPath;
 import org.carlspring.strongbox.providers.layout.p2.P2ArtifactReader;
 import org.carlspring.strongbox.storage.Storage;
 import org.carlspring.strongbox.storage.repository.Repository;
@@ -55,6 +58,30 @@ public class P2LayoutProvider
     }
 
     @Override
+    public void deleteMetadata(String storageId,
+                               String repositoryId,
+                               String metadataPath)
+        throws IOException
+    {
+        
+    }
+    
+    @Override
+    protected void doDeletePath(RepositoryPath repositoryPath,
+                                boolean force)
+        throws IOException
+    {
+        Files.delete(repositoryPath);
+
+        Repository repository = repositoryPath.getFileSystem().getRepository();
+        RepositoryFileSystemProvider provider = getProvider(repositoryPath);
+        if (force && repository.allowsForceDeletion())
+        {
+            provider.deleteTrash(repositoryPath);
+        }
+    }
+    
+    @Override
     public boolean containsArtifact(Repository repository,
                                     ArtifactCoordinates coordinates)
             throws IOException
@@ -88,84 +115,4 @@ public class P2LayoutProvider
         return containsArtifact(repository, P2ArtifactCoordinates.create(path));
     }
 
-    @Override
-    public void copy(String srcStorageId,
-                     String srcRepositoryId,
-                     String destStorageId,
-                     String destRepositoryId,
-                     String path)
-            throws IOException
-    {
-
-
-    }
-
-    @Override
-    public void move(String srcStorageId,
-                     String srcRepositoryId,
-                     String destStorageId,
-                     String destRepositoryId,
-                     String path)
-            throws IOException
-    {
-
-    }
-
-    @Override
-    public void delete(String storageId,
-                       String repositoryId,
-                       String path,
-                       boolean force)
-            throws IOException
-    {
-
-    }
-
-    @Override
-    public void deleteMetadata(String storageId,
-                               String repositoryId,
-                               String metadataPath)
-            throws IOException
-    {
-
-    }
-
-    @Override
-    public void deleteTrash(String storageId,
-                            String repositoryId)
-            throws IOException
-    {
-
-    }
-
-    @Override
-    public void deleteTrash()
-            throws IOException
-    {
-
-    }
-
-    @Override
-    public void undelete(String storageId,
-                         String repositoryId,
-                         String path)
-            throws IOException
-    {
-
-    }
-
-    @Override
-    public void undeleteTrash(String storageId,
-                              String repositoryId)
-            throws IOException
-    {
-
-    }
-
-    @Override
-    public void undeleteTrash()
-            throws IOException
-    {
-
-    }
 }
